Accept double values when setting a budget with updateBudget

+++ src/main/java/cashleh/parser/Parser.java
@@ -311,9 +311,9 @@ public class Parser {
             throw new CashLehParsingException("Ayo, type in your budget amount hor!");
         }
         String newBudget = input.split(" ", 2)[1];
-        int newBudgetAmount;
+        double newBudgetAmount;
         try {
-            newBudgetAmount = Integer.parseInt(newBudget);
+            newBudgetAmount = Double.parseDouble(newBudget);
         } catch (NumberFormatException e) {
             throw new CashLehParsingException("Eh, that's not the kind of number we flaunt in CashLeh!");
         }
